Módulo doctor

+++ authentication/views.py
@@ -5,43 +5,46 @@ from django.contrib.auth import authenticate, login, logout
 from principal.models import User
 from rest_framework.authtoken.models import Token
 from django.shortcuts import get_object_or_404
-
+from .forms import LoginForm
 
 # Create your views here.
 
+
 def login_form(request):
-    return render(request, 'login.html')
-
-
-def account(request):
-    response = HttpResponse('Login Error')
-    response = redirect(to="inicio")
 
     if request.method == 'POST':
-        username = request.POST.get('username')
-        password = request.POST.get('password')
-       
-        user = authenticate(request, username=username, password=password)
-        
-        if user is not None:
-            login(request, user) 
-            try: 
-                if get_object_or_404(Token, user=user):
-                    tk = get_object_or_404(Token, user=user)
-                    tk.delete() 
-            except:
-               pass
-
-            token = Token.objects.create(user=user)
-            response = HttpResponse('Login Sucessfully')
-            response = redirect(to=type_user(user))
-            response.set_cookie('cognitya', token)
-            if(request.POST.get('save_session')):
-                token.user.save_session = True
-                token.user.save()
-
+        login_form = LoginForm(request.POST,"form")
+        if login_form.is_valid():
+            response = HttpResponse('Login Error')
+            response = redirect(to="inicio")
+            username = login_form.cleaned_data['username']
+            password = login_form.cleaned_data['password']
+         
+            user = authenticate(request, username=username, password=password)
             
-    return response
+            if user is not None:
+                login(request, user) 
+                try: 
+                    if get_object_or_404(Token, user=user):
+                        tk = get_object_or_404(Token, user=user)
+                        tk.delete() 
+                except:
+                    pass
+
+                token = Token.objects.create(user=user)
+                response = HttpResponse('Login Sucessfully')
+                response = redirect(to=type_user(user))
+                response.set_cookie('cognitya', token)
+                if(request.POST.get('save_session')):
+                    token.user.save_session = True
+                    token.user.save()
+            return response
+    else:
+        login_form = LoginForm()
+    
+    context = {'form': login_form}
+        
+    return render(request, 'login.html', context)
 
 def type_user(user):
     if user.is_patient:
